Created the Connected Component
