+++ packages/newspringchurchapp/src/prayer/PrayerCard/flagPrayerRequest.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 export default gql`
-  mutation flagRequest($parsedId: String!) {
+  mutation flagPrayerRequest($parsedId: String!) {
     flagRequest(id: $parsedId) {
       id
     }
